requirements-colab.txt
torch
torchvision
pandas
numpy
scikit-learn
tqdm
pillow
matplotlib
requests